+++ modules/tinymce/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Fixed
 - An element could be dropped onto the decendants of a noneditable element. #TINY-9364
 
+### Improved
+- Direct invalid child text nodes of list elements will be wrapped in list item elements. #TINY-4818
+
 ## 6.3.0 - 2022-11-23
 
 ### Added